+++ rafle_lib/traits/participant_management.rs
@@ -3,10 +3,8 @@ use openbrush::contracts::access_control::AccessControlError;
 use openbrush::traits::AccountId;
 use openbrush::traits::Balance;
 
-/*
 #[openbrush::wrapper]
 pub type ParticipantManagementRef = dyn ParticipantManagement;
-*/
 
 #[openbrush::trait_definition]
 pub trait ParticipantManagement {
@@ -16,10 +14,10 @@ pub trait ParticipantManagement {
     /// weight can represent the number of raffle tickets for this participant.
     /// weight can also represent the amount staked in dAppStaking, ...
     #[ink(message)]
-    fn add_participant(&mut self, era: u128, participant: AccountId, weight: Balance) -> Result<(), ParticipantManagementError>;
+    fn add_participant(&mut self, era: u32, participant: AccountId, weight: Balance) -> Result<(), ParticipantManagementError>;
 
     /// list all participant for a given era
-    fn _list_participants(&self, era: u128) -> Vec<(AccountId, Balance)>;
+    fn _list_participants(&self, era: u32) -> Vec<(AccountId, Balance)>;
 
 }
